+++ src/main/scala/com/exasol/cloudetl/bucket/AzureBlobBucket.scala
@@ -1,11 +1,15 @@
 package com.exasol.cloudetl.bucket
 
+import scala.util.matching.Regex
+
 import com.exasol.cloudetl.storage.StorageProperties
 
 import org.apache.hadoop.conf.Configuration
 
 /** A [[Bucket]] implementation for the Azure Blob Storage */
-final case class AzureBlobBucket(path: String, params: StorageProperties) extends Bucket {
+final case class AzureBlobBucket(path: String, params: StorageProperties)
+    extends Bucket
+    with SecureBucket {
 
   private[this] val AZURE_ACCOUNT_NAME: String = "AZURE_ACCOUNT_NAME"
   private[this] val AZURE_CONTAINER_NAME: String = "AZURE_CONTAINER_NAME"
@@ -22,22 +26,16 @@ final case class AzureBlobBucket(path: String, params: StorageProperties) extend
    * Returns the list of required property keys for Azure Blob Storage.
    */
   override def getRequiredProperties(): Seq[String] =
-    Seq(AZURE_ACCOUNT_NAME)
+    Seq.empty[String]
 
-  /**
-   * Validation method specific to the Azure Blob Storage.
-   *
-   * Validates required properties by calling parent {@code validate}
-   * method. Furthermore, checks whether either [[AZURE_SECRET_KEY]] or
-   * [[AZURE_SAS_TOKEN]] parameters are available.
-   */
-  private[this] def validateExtra(): Unit = {
-    validate()
-    if (!properties.containsKey(AZURE_SECRET_KEY) && !properties.containsKey(AZURE_SAS_TOKEN)) {
-      throw new IllegalArgumentException(
-        s"Please provide a value for either $AZURE_SECRET_KEY or $AZURE_SAS_TOKEN!"
-      )
-    }
+  /** @inheritdoc */
+  override def getSecureProperties(): Seq[String] =
+    Seq(AZURE_SECRET_KEY, AZURE_SAS_TOKEN)
+
+  /** @inheritdoc */
+  override def validate(): Unit = {
+    validateRequiredProperties()
+    validateConnectionProperties()
   }
 
   /**
@@ -47,7 +45,7 @@ final case class AzureBlobBucket(path: String, params: StorageProperties) extend
    * in order to create a configuration.
    */
   override def getConfiguration(): Configuration = {
-    validateExtra()
+    validate()
 
     val conf = new Configuration()
     conf.set("fs.azure", classOf[org.apache.hadoop.fs.azure.NativeAzureFileSystem].getName)
@@ -59,17 +57,39 @@ final case class AzureBlobBucket(path: String, params: StorageProperties) extend
       classOf[org.apache.hadoop.fs.azure.Wasbs].getName
     )
 
-    val accountName = properties.getString(AZURE_ACCOUNT_NAME)
-    if (properties.containsKey(AZURE_SAS_TOKEN)) {
-      val sasToken = properties.getString(AZURE_SAS_TOKEN)
-      val containerName = properties.getString(AZURE_CONTAINER_NAME)
+    val mergedProperties = if (properties.hasNamedConnection()) {
+      properties.merge(AZURE_ACCOUNT_NAME)
+    } else {
+      properties
+    }
+
+    val accountAndContainer = regexParsePath(path)
+
+    val accountName =
+      mergedProperties.get(AZURE_ACCOUNT_NAME).getOrElse(accountAndContainer.accountName)
+
+    if (mergedProperties.containsKey(AZURE_SAS_TOKEN)) {
+      val sasToken = mergedProperties.getString(AZURE_SAS_TOKEN)
+      val containerName =
+        mergedProperties.get(AZURE_CONTAINER_NAME).getOrElse(accountAndContainer.containerName)
       conf.set(s"fs.azure.sas.$containerName.$accountName.blob.core.windows.net", sasToken)
     } else {
-      val secretKey = properties.getString(AZURE_SECRET_KEY)
+      val secretKey = mergedProperties.getString(AZURE_SECRET_KEY)
       conf.set(s"fs.azure.account.key.$accountName.blob.core.windows.net", secretKey)
     }
 
     conf
   }
+
+  private[this] final val AZURE_BLOB_PATH_REGEX: Regex =
+    """wasbs?://(.*)@([^.]+).blob.core.windows.net/(.*)$""".r
+
+  private[this] def regexParsePath(path: String): AccountAndContainer = path match {
+    case AZURE_BLOB_PATH_REGEX(containerName, accountName, _) =>
+      AccountAndContainer(accountName, containerName)
+    case _ => throw new IllegalArgumentException(s"Invalid Azure blob wasb(s) path: $path!")
+  }
+
+  private[this] case class AccountAndContainer(accountName: String, containerName: String)
 
 }