+++ src/main/scala/com/exasol/cloudetl/storage/StorageProperties.scala
@@ -2,6 +2,8 @@ package com.exasol.cloudetl.storage
 
 import java.net.URI
 
+import com.exasol.ExaConnectionInformation
+import com.exasol.ExaMetadata
 import com.exasol.cloudetl.common.AbstractProperties
 import com.exasol.cloudetl.common.CommonProperties
 
@@ -11,8 +13,10 @@ import com.exasol.cloudetl.common.CommonProperties
  * provided key-value parameters for storage import and export
  * user-defined-functions (udfs).
  */
-class StorageProperties(private val properties: Map[String, String])
-    extends AbstractProperties(properties) {
+class StorageProperties(
+  private val properties: Map[String, String],
+  private val exaMetadata: Option[ExaMetadata]
+) extends AbstractProperties(properties) {
 
   import StorageProperties._
 
@@ -47,6 +51,62 @@ class StorageProperties(private val properties: Map[String, String])
     get(PARALLELISM).fold(defaultValue)(identity)
 
   /**
+   * Returns an Exasol [[ExaConnectionInformation]] named connection
+   * information.
+   */
+  @SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
+  final def getConnectionInformation(): ExaConnectionInformation =
+    exaMetadata.fold {
+      throw new IllegalArgumentException("Exasol metadata is None!")
+    }(_.getConnection(getString(CONNECTION_NAME)))
+
+  /** Checks if the Exasol named connection property is provided. */
+  final def hasNamedConnection(): Boolean =
+    containsKey(CONNECTION_NAME)
+
+  /**
+   * Returns a new [[StorageProperties]] what merges the key-value pairs
+   * parsed from user provided Exasol named connection object.
+   */
+  final def merge(accountName: String): StorageProperties = {
+    val connectionParsedMap = parseConnectionInfo(accountName)
+    val newProperties = properties ++ connectionParsedMap
+    new StorageProperties(newProperties, exaMetadata)
+  }
+
+  /**
+   * Parses the connection object password into key-value map pairs.
+   *
+   * If the connection object contains the username, it is mapped to the
+   * {@code keyForUsername} parameter. However, this value is
+   * overwritted if the provided key is available in password string of
+   * connection object.
+   */
+  private[this] def parseConnectionInfo(keyForUsername: String): Map[String, String] = {
+    val connection = getConnectionInformation()
+    val username = connection.getUser()
+    val password = connection.getPassword();
+    val map = password
+      .split(";")
+      .map { str =>
+        val idx = str.indexOf('=')
+        if (idx < 0) {
+          throw new IllegalArgumentException(
+            "Connection object password does not contain key=value pairs!"
+          )
+        }
+        str.substring(0, idx) -> str.substring(idx + 1)
+      }
+      .toMap
+
+    if (username.isEmpty()) {
+      map
+    } else {
+      Map(keyForUsername -> username) ++ map
+    }
+  }
+
+  /**
    * Returns a string value of key-value property pairs.
    *
    * The returned string is sorted by keys ordering.
@@ -72,9 +132,26 @@ object StorageProperties extends CommonProperties {
   /** An optional property key name for the parallelism. */
   private[storage] final val PARALLELISM: String = "PARALLELISM"
 
-  /** Returns [[StorageProperties]] from key-value pairs map. */
+  /** An optional property key name for the named connection object. */
+  private[storage] final val CONNECTION_NAME: String = "CONNECTION_NAME"
+
+  /**
+   * Returns [[StorageProperties]] from key values map and
+   * [[ExaMetadata]] metadata object.
+   */
+  def apply(params: Map[String, String], metadata: ExaMetadata): StorageProperties =
+    new StorageProperties(params, Option(metadata))
+
+  /** Returns [[StorageProperties]] from only key-value pairs map. */
   def apply(params: Map[String, String]): StorageProperties =
-    new StorageProperties(params)
+    new StorageProperties(params, None)
+
+  /**
+   * Returns [[StorageProperties]] from properly separated string and
+   * [[ExaMetadata]] metadata object.
+   */
+  def apply(string: String, metadata: ExaMetadata): StorageProperties =
+    apply(mapFromString(string), metadata)
 
   /** Returns [[StorageProperties]] from properly separated string. */
   def apply(string: String): StorageProperties =